(Dyning Details) 쿼리

+++ phoncnic/src/main/java/ds/com/phoncnic/entity/Follow.java
@@ -4,6 +4,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
 import lombok.AllArgsConstructor;
@@ -19,15 +20,16 @@ import lombok.ToString;
 @Getter
 @ToString
 public class Follow {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long fno;
+    private int fno;
 
     @ManyToOne
     private Member follower;
-    
+
     private String dyningname;
     private String artistname;
 
-    
+
 }
